backend/requirements.txt
# Core Dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Development Tools
django-debug-toolbar==4.2.0
black==23.11.0
flake8==6.1.0

# API Documentation
drf-yasg==1.21.7

# Security
django-defender==0.9.7
django-axes==6.1.1

# Utilities
Pillow==10.1.0
celery==5.3.4
redis==5.0.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

django-extensions==3.2.3
Werkzeug==3.1.3
pyOpenSSL==24.3.0
argon2-cffi==23.1.0

django-otp==1.5.4
djangorestframework-simplejwt==5.3.1